Fix tempo
While I was testing I found some issues with the slower tempos, I've implemented a fix to those issues.

+++ VR_Piano/Assets/Scripts/SelectedSongManager.cs
@@ -8,7 +8,7 @@ public class PlaySceneManager : MonoBehaviour
     private int selectedSongIndex;
     private bool left_enabled;
     private bool right_enabled;
-    private int tempo_multiplier;
+    private float tempo_multiplier;
     [SerializeField] private MidiMessages midiMessages; // create reference to MidiMesseges component
 
     void Start()
@@ -20,13 +20,13 @@ public class PlaySceneManager : MonoBehaviour
         // Initialize the other parameters
         left_enabled = PlayerPrefs.GetInt("LeftEnabled", 1) == 1;
         right_enabled = PlayerPrefs.GetInt("RightEnabled", 1) == 1;
-        tempo_multiplier = PlayerPrefs.GetInt("TempoMultiplier", 1);
+        tempo_multiplier = PlayerPrefs.GetFloat("TempoMultiplier", 1);
 
         // Call function to play the selected song
         PlaySelectedSong();
     }
 
-    private async void  PlaySelectedSong()
+    private async void PlaySelectedSong()
     {
         // Checks that MidiMessages component is assigned
         if (midiMessages != null)
